@@ -1,7 +1,10 @@
 module github.com/leonklingele/grouper
 
-go 1.18
+go 1.24.2
 
-require golang.org/x/tools v0.19.0
+require golang.org/x/tools v0.33.0
 
-require golang.org/x/mod v0.16.0 // indirect
+require (
+	golang.org/x/mod v0.24.0 // indirect
+	golang.org/x/sync v0.14.0 // indirect
+)
